requirements.txt
# Core dependencies
flask==2.3.3
werkzeug==2.3.7
sqlalchemy==1.4.50
alembic==1.11.3

# Face Recognition
cmake>=3.25.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
opencv-python==4.8.0.76
numpy>=1.24.0,<2.0.0

# Utils
pillow==10.2.0
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.3.0
pyasn1>=0.4.1,<0.5.0
pyasn1-modules>=0.2.8,<0.3.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4 